+ finished the underlying code for scanning a port + cafebabe can now scan a single port by using the -p and providing   the port u wish to scan + started on work multi scan code, may have to make some changes to   the pre-existing "framework"
Changes to be committed:
	modified:   include/bpf.c
	modified:   include/bpf.h
	modified:   include/cafebabe.c
	modified:   include/framez.c
	modified:   include/framez.h
	modified:   include/scan.c
	modified:   include/scan.h
	modified:   main.c

+++ src/include/framez.c
@@ -16,7 +16,7 @@ unsigned short checksum(unsigned short *p, int l){
 }
 
 /* does not need an explanation */
-void ipv4Hdr(char *b, unsigned short l, unsigned short id, unsigned short off, unsigned char ttl, unsigned char p, unsigned short sum, unsigned long src, unsigned long dst){
+void ipv4Hdr(char *b, unsigned short l, unsigned short id, unsigned short off, unsigned char ttl, unsigned char p, unsigned short sum, unsigned int src, unsigned int dst){
     struct ip *iphdr = (struct ip *)(b);
     iphdr->ip_hl = 5;
     iphdr->ip_v = 4;
